refactor: rename version info to dependency info
