+++ apps/ensindexer/src/lib/dependency-info.ts
